Remove hiera override of enabled_import_methods for Glance
I don't want to use glance-direct without a shared filesystem for staging.
The right way to do it is not with hiera but with GlanceEnabledImportMethods.

I need to override GlanceEnabledImportMethods for updating central or I'll only
get web download and not copy-image:

[root@control-plane-controller-1 ~]# podman exec glance_api cat /etc/glance/glance-api.conf | grep -v \# | grep import_methods
enabled_import_methods=[web-download]
[root@control-plane-controller-1 ~]#

but for dcn copy-image gets enabled via dcn-hci.yaml.

 https://github.com/openstack/tripleo-heat-templates/blob/master/environments/dcn-hci.yaml#L27