Update standard deploy.sh
As tripleo-client will now run ansible from the start and always
create config-download in /home/stack/ with a script to trigger
it later (as per the removal of mistral) let's switch to the new
default.

Use the HEAT as usual and it will include what DOWN and CONFIG
used to do by default.

If there's a need to re-run only a part of config-download, then
use DOWN to only run the config-download ansible. Examples are
provided of some of my favorite useful tags.