Remove problematic ANSIBLE_SSH_CONTROL_PATH override
I had an old personal override of ANSIBLE_SSH_CONTROL_PATH [1].
It casued very strange behavior when combined with the new
ANSIBLE_SSH_CONTROL_PATH_DIR as set by tripleo-ansible [2].
Setting both of these vars is odd especially when they are
set to different paths as per a recent update [2].

The behavior it produced was even stranger. Ansible was confusing
its facts for hosts and I was getting mon containers deployed on
hosts in the [osds] inventory even though the [mons] group in the
inventory was correct. Kind of frustrating that it wasted my time
and others.

[1] https://docs.ansible.com/ansible/latest/reference_appendices/config.html#ansible-ssh-control-path
[2] https://github.com/openstack/tripleo-ansible/commit/ca165ec94018293dfbbbd37c44a2704b8144f203

                __________                  __,___/  ",",
         ___.--"          "\'.         ____/  l   \    ",'-,
  ------f"               // \\\        \  (l\ \    \     \ \",
        |                    |||       /   u       |      \ \ \
        |                    |||     _ )          /       | |  \
    ----L_-XXX-.             .|'    / U          <        | |  |
                "\   -<_____///     \           6 )       | |  |
                  \___)     -"       '.       -.<"       / /   |
                                      |'.___  |       _._."   /
                                      |     ./     _."."   _."
                                     /      |"----"     _."
                                  jjs       \